block: use "unsigned long" for blk_validate_block_size() (git-fixes).
suse-commit: 3e9ef51b9e48a722274cfd7e6e55b4dd30836048

+++ include/linux/blkdev.h
@@ -308,7 +308,7 @@ static inline unsigned short req_get_ioprio(struct request *req)
 	return req->ioprio;
 }
 
-static inline int blk_validate_block_size(unsigned int bsize)
+static inline int blk_validate_block_size(unsigned long bsize)
 {
 	if (bsize < 512 || bsize > PAGE_SIZE || !is_power_of_2(bsize))
 		return -EINVAL;
